+++ packages/backend-platform-test-stubs/src/import_path_verifier_stub.ts
@@ -1,0 +1,13 @@
+import { ImportPathVerifier } from '@aws-amplify/plugin-types';
+
+/**
+ * Stub implementation of ImportPathVerifier
+ */
+export class ImportPathVerifierStub implements ImportPathVerifier {
+  /**
+   * @inheritDoc
+   */
+  verify = (): void => {
+    // noop
+  };
+}