+++ packages/integration-tests/src/process-controller/predicated_action_queue_builder.ts
@@ -87,7 +87,8 @@ export class PredicatedActionBuilder {
     this.getLastPredicatedAction().then = {
       actionType: ActionType.ASSERT_ON_PROCESS_OUTPUT,
       action: (strWithDeploymentTime: string) => {
-        const regex = /^✨ {2}Total time: (\d*.\d*).*$/;
+        // the time can be in fractional or whole seconds. 24.3, 24, 24.22 etc.
+        const regex = /^✨ {2}Total time: (\d*\.*\d*)s.*$/;
         const deploymentTime = strWithDeploymentTime.match(regex);
         if (
           deploymentTime &&